+++ packages/transformer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ts-linq/transformer
 
+## 2.1.31
+
+### Patch Changes
+
+- Updated dependencies
+  - @ts-linq/types@4.10.0
+  - @ts-linq/ast@3.2.10
+
 ## 2.1.30
 
 ### Patch Changes